Initial commit, i.e. hello world
Get deadbeef in a register and get it running under RISC-V qemu.
Set up the project directory structure, recursive Make, the basics
of GDB and all the necessary cflags to get going. Write a low-level
assembly routine to kick things off, and a linker script to place
it properly.

Respect the Plan 9 directory hierarchy convention, and to a degree
the 9front coding style. TODO: write a perl script to auto-do this
as things get more complicated.

Signed-off-by: Jay Lang <jaytlang@mit.edu>