Implement the page-grained memory allocator
Along with a slew of tweaks to get everything to compile properly,
implement the page-grained allocator in msys. Additionally, create
a means to panic the kernel (through ultimateyeet() in lsys), and
a library subsystem that theoretically both users and the kernel
can link to. This is the beginnings of a libc (though the panic routine
should probably hit the hardware abstraction layer...)

Signed-off-by: Jay Lang <jaytlang@mit.edu>