Write logic for kernel virtual memory, and enable it + nits
Enhance msys significantly to provide support for virtual
memory, specifically the SV39 scheme. Write logic for messing
with both virtual addresses and page tables themselves,
along with a catch-all to map into the kernel (/arbitrary)
address spaces.

Create an initial kernel address space with the help of
linker symbols which encompasses the UART, kernel text,
and remainder of the address space.

With the help of modified hsys and startup (an mret to supervisor
mode vs. continuing in machine mode), install the new page
table and verify it works. This boundary is kinda flimsy
and might be worth a rethink down the road, but I think
the current choice is fine. Maybe shifting should occur
ahead of time, but this is more error proof.

Improve upon the standard library to introduce round up/
round down operations - like ceil() and floor() but for the
specific integer-only use case. Doesn't hurt to have a more
complete standard library, and this is particularly useful for
msys (where it would otherwise use a macro definition etc.)

Nits: set etext to not include the rodata section in the linker
script, and incorporate the ekernel symbol for msys code.

Signed-off-by: Jay Lang <jaytlang@mit.edu>