Add a user trap vector and map it properly. Write process data structs.
A trampoline page is set up at the top of the virtual address space -
doubly mapped in kernel mode and to-be mapped in user mode as well.
A trap frame is going to wind up below that page, which'll be written
into sscratch for load by the trap vector before we jump into kernel
mode.

Speaking of, the process data structure is written here as well. Add
fields for fields, using the global process lock rather than its
own just to make sure that everything works nice. Context for the
kernel is drafted (i.e. scheduler context, per-process context),
under the assumption of a single processor. SMP will come later.

Start of a process allocation routine is written, just to get an
idea. I detoured after that, and decided that a commit after writing
the trap vector was fine haha

Signed-off-by: Jay Lang <jaytlang@mit.edu>