Add process creation logic, without some trapret-specific details
Make processes, and adjust the msys API to be as minimal as possible
while allowing all the functionality we need. Carefully examine locking
discipline during the mkproc call so that we don't race or deadlock.
If something runs out of memory, panic immediately to not need to worry
about the consequences.

Signed-off-by: Jay Lang <jaytlang@mit.edu>