Export lock count to other subsystems
As I mentioned, this is a particularly useful feature for someone
like psys when performing a context switch. Even though it probably
doesn't matter all that much, this could be useful for other stuff
too.

Test and verify that this works, which it does, along with some
potential failure cases. Locks in asys are NOT re-entrant.

Signed-off-by: Jay Lang <jaytlang@mit.edu>