Export oldintrstate to allow for the process lock to work
The process lock may not be 300% necessary right now, but is
still wanted to prevent any nasty race conditions. This is a
proof of concept; coarse-grained locking isn't horrible and
will probably be helpful to diagnose any problems with threading.
Basically this is the monitor pattern.

The scheduler, when written, will do stuff WHILE INTERRUPTIBLE.
Then it'll acquire the lock on the process to switch to (thus
killing interrupts) and context switch. This means the process
lock is held during the context switch, and it's expected to
be held again by the time we return -- however, the state of
that kernel thread BEFORE the lock is acquired in terms of
interruptibility is not known and could go either way (say,
on in a syscall or off during a timer interrupt).

Resolve this by exporting oldintrstate and allowing get/set
operations on it via the hsys api. Hopefully these issues become
a little bit more concrete as I continue on.

Signed-off-by: Jay Lang <jaytlang@mit.edu>