Write the OS scheduler + yield coroutines for kernel threading
Kernel threading is now enabled via yield, llcontextswitch,
and the scheduler itself. A global context is configured on the
kernel stack, which is still shared amongst all process contexts.
This will change if things get messy.

In the meantime, the linked list of processes is cycled through
in round-robin fashion, and as we switch over to the new
context we hold onto the process lock to avoid weird interrupt
interleavings (since the scheduler is pre-emptible, even though
interrupts are off during the switch). Yield goes back the other
way, particularly restoring the proper interrupt state for its
context since that's the only thing llcontextswitch() doesn't
get. There's a pretty good comment about this in hsys/spinlock.c
if my memory serves me.

Still to do on the process side before testing can begin: write
the rest of the user trap handler (utrap2), and set up the initial
process table + return routines, including the drop to user mode.

Signed-off-by: Jay Lang <jaytlang@mit.edu>