Write user trap handler + return mechanisms
Write the user trap handler in both assembly and C,
from high trampoline page to the kernel and back
again. Right now handle only timer interrupts, but that's
fine; still log all of the interrupts in the console w/ dsys.

This requires considerable help from each of the subsystems
and some interesting prototypes - hence why I stashed and
did the clang-format-ing.

All there is left to do is finalize process creation routines
and ensure that we can populate an initial process table,
before dropping to userland and checking that code runs +
can get pre-empted and returned to. I should also assert
that utrap.S doesn't exceed a page.

We're almost done with this thing and almost ready to start
debugging. Cheers.

Signed-off-by: Jay Lang <jaytlang@mit.edu>