Tweak the linker script to enforce a one page utrap
The assert statement added ensures that we only need
to map one page to cover the trap page, and it occurs at a well
known location we're able to use in msys etc.

Additionally, create a map as part of the build process to
inspect build artifacts and verify that everything works as it
should, including page-sized paddings and library linkings (which
happen to hit rodata).

Signed-off-by: Jay Lang <jaytlang@mit.edu>