Nits: remove kstack from psys, as it's not needed
Everything has a global kernel stack and operates on it - so it's
not necessary to store it anywhere besides in the top level traps.
If we were switching stacks per process, this would be necessary,
but it isn't (for now, pending any stack overflow issues).

As a result, delete the kstack and have procinit be passed nothing.

Signed-off-by: Jay Lang <jaytlang@mit.edu>