Add user memory allocation routine, in preparation for exec/frun
Get a user memory allocation routine, and export a set of defines
that other subsystems (psys) can use to specify whether a page
holds code or data. I'll need to go through the linker for
those code pages, but shouldn't be an issue (besides some tedium)
to generate the right script.

It's entirely possible I could perl that out...in the meantime,
grab memory from allocpage() and panic at OOM as usual. Map it in
directly at the size passed, which makes it v compatible with
currentproc->size.

Signed-off-by: Jay Lang <jaytlang@mit.edu>