Create srun: a fork/exec analogue to get off the ground
srun can be invoked regardless of a process context, and utilizes
the machinery we've built so far to inject a new process into the
scheduler. srun copies code onto the heap with the help of a new
walk primitive in msys and mkuserpages, which takes a code
corresponding to the types of protections msys should apply at the
page level. srun also takes the start/end of a given set of functions
to copy in, which can be obtained through careful use of the linker.

In the future, this approach (copy everything in) is not dissimilar
to how an ELF loader works. The block driver will probably copy
something into memory and ship it to us to work with if we
ever get that far.

This required a few new helpers to get done, so implement memcpy
in lsys and utilize it for copying of code.

Nits: ignore .swp files for vi.

Signed-off-by: Jay Lang <jaytlang@mit.edu>