Finalize the pre-emptive scheduler
Reorder the Makefile and avoid circular dependencies through
a top-level crossmod.c to cross interfaces already relied upon
in the other direction.

Write user mode test code and inject it directly into the linker
script. Tweak the init log to reflect this, along with other changes
to the system scheduler etc.

Stop using la and export functions to obtain parts of the trap
frame. Use virtual addressing instead to avoid page faults. To this
end, reorder much of utrap to properly respect the page table
flip-flop, and introduce a setstvec routine in the kernel to restore
/set in-kernel trap handlers for pre-emption at the right time.
Also make sure that the kernel page table is not loaded in through
la -- rather, add it to the trap frame and access it in utrap
to avoid a page fault.

Per-process kernel stacks now exist to permit full context
switching from the scheduler's global context into process
specific stacks. This is critical to allow the stack to reset upon
each re-entry to the kernel, which it now can do without bugs.
Enable this and patch a particularly nasty bug that involved adding
PAGESIZE to sp at two different times, overwriting nearby pages.

Register the global context kernel stack so that we know how to
switch to it without issue from other subsystems. Additionally,
for per-process kernel stacks, add the relevant proc field and
set up the constant trap frame stack setting to allow for
stack resetting to the base, and not continually building on it.

Preserve interrupt state in the scheduler, ensuring it is
interruptible by the hardware. Empirically this probably works.

Lastly, correct a bug in the process creation routine that failed to
recognize starting program counter addresses within the first
page or so, and miscalculated where they actually were. This
shouldn't be an issue once an initrd is included anyway, but
is helpful for test.

This marks the first release of asys that's capable of running
processes. Hence I'll tag it 0.0.1 shortly. Rejoice!

Signed-off-by: Jay Lang <jaytlang@mit.edu>