Add support for system calls
Complete initial work towards system call suites, and test via
user.S. In practice, this is pretty simple: add a new hook into
psys for inspecting the contents of the trapframe, advance the
pc so we don't repeatedly ecall, and catch it in the first place
via the right scause masks.

Next: building out the syscall suite.

Signed-off-by: Jay Lang <jaytlang@mit.edu>