Write the initial ramdisk + prepare for serious userland
Write a very basic initial ramdisk implementation, which has
a magic number for consistency, and pointers to file size and
offset within the image + the names themselves. At compile time,
this initrd is generated via the contents of the initrd/progs
directory in the source, and then linked into a relocatable
object file with riscv64-unknown-elf-ld.

The old method of running files from directly within the kernel
binary is still possible, and should work with some less sophisticated
C files? I am going to attempt to write psrv (exposes procfs) and memsrv
(exposes the initrd) using this method to avoid the chicken-and-egg
problem I've been facing.

Signed-off-by: Jay Lang <jaytlang@mit.edu>