Refine syscall convention, fix trapframe->sp bug
Write a more complex syscall frontend capable of handling
different syscalls, and establish conventions for dealing
with switching in and out of the syscall context: enable
interrupts, increment epc, release the proc lock, etc.

Write a test program that ensures this harness is mostly
AOK.

Nits: fix a bug in psys which for some reason set the trapframe
sp to point to its physical address instead of its virtual
address. This caused an immediate page fault when the stack
was stored to...fix this by using the proper virtual address
(in otherwords, prgrmsz + 2 * PAGESIZE).

Nits: exclude initrd output products from the git repository in
the .gitignore

Signed-off-by: Jay Lang <jaytlang@mit.edu>