Refine message passing
Pass message using a source, sequence number, and message ONLY.
Because of the relatively small number of messages at play here
(think 9P in scope), this is completely fine and way hecking easier
than how things were. Memory grant type functionality will still
definitely be necessary, but we can totally back burner it in favor
of macros and other niceities for the time being -- ESPECIALLY
for the bootstrap environment.

Signed-off-by: Jay Lang <jaytlang@mit.edu>